# docs/api/openapi.yaml
openapi: 3.0.0
paths:
  /auth/login:
    post:
      description: >-
        Returns the same 15-minute access token and 7-day refresh token pair as
        OTP verification.
      operationId: Auth_login
      parameters: []
      requestBody:
        required: true
        description: Logger phone number and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
            examples:
              logger:
                summary: Configured logger user
                value:
                  phone: '+212600000001'
                  password: zomaal01@
      responses:
        '200':
          description: Authentication succeeded.
          headers:
            Cache-Control:
              description: Prevents access and refresh tokens from being cached.
              schema:
                type: string
                example: no-store
            Pragma:
              description: Legacy cache prevention for token responses.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponseDto'
        '400':
          description: Invalid or unexpected request field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '401':
          description: Invalid phone number or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '429':
          description: More than 5 login attempts in 15 minutes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      summary: Log in with phone number and password
      tags:
        - Auth
  /auth/send-otp:
    post:
      description: >-
        Requires `Content-Type: application/json`. Requests a six-digit OTP
        through SMS or WhatsApp. This endpoint does not require an access token
        and is limited to three requests per phone/channel in ten minutes.
      operationId: Auth_sendOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpDto'
            examples:
              sms:
                summary: Send by SMS
                value:
                  phone: '+212612345678'
                  channel: sms
              whatsapp:
                summary: Send by WhatsApp
                value:
                  phone: '+212612345678'
                  channel: whatsapp
      responses:
        '200':
          description: >-
            OTP request accepted. In development OTP mode no SMS is sent; use
            DEV_OTP_CODE.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseDto'
        '400':
          description: Invalid phone number, channel, or unexpected request field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '429':
          description: More than 3 OTP requests in 10 minutes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: OTP provider is not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      summary: Send OTP to phone number
      tags:
        - Auth
  /auth/verify-otp:
    post:
      description: >-
        Creates the user on first successful verification. Returns a 15-minute
        access token and a 7-day refresh token.
      operationId: Auth_verifyOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpDto'
            examples:
              default:
                summary: Verify a six-digit OTP
                value:
                  phone: '+212612345678'
                  otp: '123456'
      responses:
        '200':
          description: Returns access token, refresh token, and profile status
          headers:
            Cache-Control:
              description: Prevents access and refresh tokens from being cached.
              schema:
                type: string
                example: no-store
            Pragma:
              description: Legacy cache prevention for token responses.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponseDto'
        '400':
          description: Invalid input or invalid/expired OTP.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '429':
          description: More than 5 verification attempts in 10 minutes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      summary: Verify OTP and log in or register
      tags:
        - Auth
  /auth/refresh:
    post:
      description: >-
        The supplied refresh token is single-use after rotation. Store the newly
        returned refresh token.
      operationId: Auth_refreshTokens
      parameters: []
      requestBody:
        required: true
        description: >-
          Send the refresh token in the JSON body, not in the Authorization
          header.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
            examples:
              default:
                summary: Rotate the refresh token
                value:
                  refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: New access and refresh tokens.
          headers:
            Cache-Control:
              description: Prevents access and refresh tokens from being cached.
              schema:
                type: string
                example: no-store
            Pragma:
              description: Legacy cache prevention for token responses.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairDto'
        '400':
          description: Missing, empty, or unexpected request field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '401':
          description: Refresh token is invalid, expired, revoked, or not a refresh token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      summary: Rotate refresh token and issue a new token pair
      tags:
        - Auth
...
  /stores:
    post:
      description: >-
        Creates one store for the authenticated user and marks onboarding as
        complete.
      operationId: Stores_create
      parameters: []
      requestBody:
        required: true
        description: Store profile and pickup address for the current user.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreDto'
            examples:
              moroccoStore:
                summary: Moroccan online store
                value:
                  ownerName: Ahmed Alaoui
                  businessName: Atlas Market
                  address: 123 Rue Hassan II
                  city: Casablanca
                  country: Morocco
                  logoUrl: https://example.com/logo.png
      responses:
        '201':
          description: Store created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponseDto'
        '400':
          description: Invalid, missing, or unexpected request field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '401':
          description: Missing or invalid Zomaal bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '409':
          description: The authenticated user already has a store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearer: []
      summary: Create the current user’s store
      tags:
        - Stores
  /stores/me:
    get:
      description: >-
        Requires `Authorization: Bearer <accessToken>`. Returns the single store
        owned by the authenticated user.
      operationId: Stores_findOne
      parameters: []
      responses:
        '200':
          description: Current user’s store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponseDto'
        '401':
          description: Missing or invalid Zomaal bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: The authenticated user has not created a store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearer: []
      summary: Get the current user’s store
      tags:
        - Stores
    put:
      description: All request fields are optional; only supplied fields are changed.
      operationId: Stores_update
      parameters: []
      requestBody:
        required: true
        description: >-
          Partial update. Include at least one field; omitted fields remain
          unchanged.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreDto'
            examples:
              renameStore:
                summary: Update business name and logo
                value:
                  businessName: Atlas Market Pro
                  logoUrl: https://example.com/new-logo.png
              updateAddress:
                summary: Update pickup address
                value:
                  address: 45 Boulevard Zerktouni
                  city: Casablanca
                  country: Morocco
      responses:
        '200':
          description: Updated store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponseDto'
        '400':
          description: Invalid or unexpected request field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '401':
          description: Missing or invalid Zomaal bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: The authenticated user has not created a store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearer: []
      summary: Update the current user’s store
      tags:
        - Stores
...
info:
  title: Zomaal API
  description: >-
    Production API reference for Zomaal authentication, stores, and shipping
    providers.


    ### Required headers

    - Requests with a JSON body: `Content-Type: application/json`

    - JSON responses: `Accept: application/json` (recommended)

    - Protected endpoints: `Authorization: Bearer <accessToken>`

    - Provider webhook signature headers are documented on each webhook
    operation.


    ### Authentication

    Use `POST /auth/login` or `POST /auth/verify-otp` to receive an access token
    and refresh token. Click **Authorize** and paste the access token only;
    Swagger UI adds the `Bearer` prefix. Send refresh tokens only in the JSON
    body of `POST /auth/refresh`.


    ### Response format

    Success responses use the schema shown per operation. Errors use `{ message,
    error, statusCode }`; validation errors can return an array in `message`.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: >-
        Paste the access token only. Do not paste the refresh token or include
        the word Bearer.
  schemas:
    LoginDto:
      type: object
      properties:
        phone:
          type: string
          description: Logger phone number in international E.164 format.
          example: '+212600000001'
          pattern: ^\+[1-9]\d{7,14}$
        password:
          type: string
          description: Logger password configured through LOGGER_PASSWORD.
          example: zomaal01@
          minLength: 8
          format: password
          writeOnly: true
      required:
        - phone
        - password
    AuthTokensResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          description: Zomaal bearer token. Expires after 15 minutes.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Token used only with POST /auth/refresh. Expires after 7 days.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        isProfileCompleted:
          type: boolean
          description: Whether the user has completed store onboarding.
          example: false
      required:
        - accessToken
        - refreshToken
        - isProfileCompleted
    ApiErrorDto:
      type: object
      properties:
        message:
          description: >-
            Human-readable error detail. Validation failures return one or more
            messages.
          oneOf:
            - type: string
              example: Invalid or expired OTP
            - type: array
              items:
                type: string
              example:
                - phone must be a valid phone number
        error:
          type: string
          description: HTTP error reason phrase.
          example: Bad Request
        statusCode:
          type: number
          description: HTTP status code.
          example: 400
      required:
        - message
        - error
        - statusCode
    SendOtpDto:
      type: object
      properties:
        phone:
          type: string
          description: User phone number in international E.164 format.
          example: '+212612345678'
          pattern: ^\+[1-9]\d{7,14}$
        channel:
          type: string
          description: OTP delivery channel.
          example: sms
          enum:
            - sms
            - whatsapp
      required:
        - phone
        - channel
    MessageResponseDto:
      type: object
      properties:
        message:
          type: string
          example: OTP sent successfully
      required:
        - message
    VerifyOtpDto:
      type: object
      properties:
        phone:
          type: string
          description: The same E.164 phone number used to request the OTP.
          example: '+212612345678'
          pattern: ^\+[1-9]\d{7,14}$
        otp:
          type: string
          description: Six-digit verification code.
          example: '123456'
          minLength: 6
          maxLength: 6
          pattern: ^\d{6}$
      required:
        - phone
        - otp
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          description: >-
            Refresh token returned by login, verify-otp, or the previous refresh
            call. Send it in the JSON body, not in the Authorization header.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - refreshToken
    TokenPairDto:
      type: object
      properties:
        accessToken:
          type: string
          description: Zomaal bearer token. Expires after 15 minutes.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Token used only with POST /auth/refresh. Expires after 7 days.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - accessToken
        - refreshToken
    StoreResponseDto:
      type: object
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        ownerName:
          type: string
          example: John Doe
        businessName:
          type: string
          example: John Electronics
        address:
          type: string
          example: 123 Main Street
        city:
          type: string
          example: Casablanca
        country:
          type: string
          example: Morocco
        logoUrl:
          type: string
          example: https://example.com/logo.png
          nullable: true
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          example: '2026-07-16T10:30:00.000Z'
          format: date-time
        updatedAt:
          type: string
          example: '2026-07-16T10:30:00.000Z'
          format: date-time
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440001
      required:
        - id
        - ownerName
        - businessName
        - address
        - city
        - country
        - logoUrl
        - isActive
        - createdAt
        - updatedAt
        - userId
...
    CreateStoreDto:
      type: object
      properties:
        ownerName:
          type: string
          description: Store owner’s full name.
          example: Ahmed Alaoui
        businessName:
          type: string
          description: Customer-facing business/store name.
          example: Atlas Market
        address:
          type: string
          description: Store or pickup address.
          example: 123 Rue Hassan II
        city:
          type: string
          example: Casablanca
        country:
          type: string
          example: Morocco
        logoUrl:
          type: string
          description: Public HTTPS URL of the store logo.
          example: https://example.com/logo.png
          format: uri
      required:
        - ownerName
        - businessName
        - address
        - city
        - country
    UpdateStoreDto:
      type: object
      properties:
        ownerName:
          type: string
          description: Store owner’s full name.
          example: Ahmed Alaoui
        businessName:
          type: string
          description: Customer-facing business/store name.
          example: Atlas Market
        address:
          type: string
          description: Store or pickup address.
          example: 123 Rue Hassan II
        city:
          type: string
          example: Casablanca
        country:
          type: string
          example: Morocco
        logoUrl:
          type: string
          description: Public HTTPS URL of the store logo.
          example: https://example.com/logo.png
          format: uri
...